# CONNECT-HUB-Api-Gateway/cmd/docs/swagger.yaml
definitions:
...
  models.RecruiterLogin:
    properties:
      email:
        type: string
      password:
        maxLength: 20
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  models.RecruiterProfile:
    properties:
      about_company:
        type: string
      company_name:
        type: string
      company_size:
        type: integer
      contact_email:
        type: string
      contact_phone_number:
        type: integer
      headquarters_address:
        type: string
      id:
        type: integer
      industry:
        type: string
      website:
        type: string
    type: object
  models.RecruiterSignUp:
    properties:
      about_company:
        type: string
      company_name:
        type: string
      company_size:
        type: integer
      confirm_password:
        type: string
      contact_email:
        type: string
      contact_phone_number:
        type: integer
      headquarters_address:
        type: string
      industry:
        type: string
      password:
        type: string
      website:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact:
    name: API Support
  description: ConnectHub is a professional networking platform.
  title: Go + Gin professional networking platform API Connect Hub
  version: 1.0.0
paths:
...
  /recruiter/login:
    post:
      consumes:
      - application/json
      description: Authenticate a recruiter
      parameters:
      - description: Recruiter credentials for login
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.RecruiterLogin'
      produces:
      - application/json
      responses:
        "200":
          description: Recruiter login successful
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Incorrect request format or missing required fields
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 'Internal server error: failed to authenticate recruiter'
          schema:
            $ref: '#/definitions/response.Response'
      summary: Recruiter login
      tags:
      - Recruiter Authentication Management
  /recruiter/policies:
    get:
      consumes:
      - application/json
      description: Retrieves all policies from the system.
      produces:
      - application/json
      responses:
        "200":
          description: All policies retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Failed to retrieve policies
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Get All Policies
      tags:
      - Recruiter Policy Management
  /recruiter/policy:
    get:
      consumes:
      - application/json
      description: Retrieves a single policy based on the provided policy ID.
      parameters:
      - description: Policy ID
        in: query
        name: policy_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Policy retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Failed to retrieve policy or incorrect data format
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Get One Policy
      tags:
      - Recruiter Policy Management
  /recruiter/profile:
    get:
      consumes:
      - application/json
      description: Retrieves the profile of the logged-in recruiter.
      produces:
      - application/json
      responses:
        "200":
          description: Profile retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Failed to retrieve profile
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Get Recruiter Profile
      tags:
      - Recruiter Profile Management
    patch:
      consumes:
      - application/json
      description: Allows a recruiter to edit their profile information.
      parameters:
      - description: Recruiter ID
        in: path
        name: id
        required: true
        type: integer
      - description: Recruiter Profile Data
        in: body
        name: recruiterData
        required: true
        schema:
          $ref: '#/definitions/models.RecruiterProfile'
      produces:
      - application/json
      responses:
        "200":
          description: Recruiter profile updated successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Failed to update profile
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Edit Recruiter Profile
      tags:
      - Recruiter Profile Management
  /recruiter/signup:
    post:
      consumes:
      - application/json
      description: Register a new recruiter
      parameters:
      - description: Recruiter signup data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.RecruiterSignUp'
      produces:
      - application/json
      responses:
        "200":
          description: Recruiter signup successful
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Incorrect request format or missing required fields
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 'Internal server error: failed to signup recruiter'
          schema:
            $ref: '#/definitions/response.Response'
      summary: Recruiter signup
      tags:
      - Recruiter Authentication Management
...
securityDefinitions:
  BearerTokenAuth:
    description: 'For accessing the API, you need to include "Jobseeker","Admin" or
      "Recruiter" before your token in the Authorization header. Example: "Authorization:
      Admin <your_token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"